backend/requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
Django==3.2.25
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
idna==3.10
joblib==1.5.1
numpy==2.2.6
pandas==2.3.1
pillow==11.3.0
PyJWT==2.10.1
pymongo==3.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
soupsieve==2.7
sqlparse==0.2.4
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
